Spec FileNameCleaner further
I suspect this test won't run on Windows if files can't be created with ':' in the first place :/

Maybe we can tag tests and skip them on Windows or something.

Could also remove the code for migrating unclean names, but I don't like the
user experience of having to redownload files due to changes in the script.

cc @andyeskridge